Enforce string type for messageId
Message sending currently fails with INVALID_JSON if messageId type is something other than string (number etc), as detailed in the FCM XMPP spec.

The resulting `nack` is not currently reported by Sender at all, as the returned `message_id` is always an empty string in the case of an invalid messageId in the original message; this error fails silently, and no Result is returned. Strict type enforcement would prevent such attempts to send invalid messages, and inform the user about the problem.